+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for bitcoin-scripting 
 
+## 0.2.0
+
+* Refactors script descriptors module in order to make invalid descriptors unrepresentable.
+* Adds utility functions to compile script descriptors and produce addresses from output descriptors.
+
 ## 0.1.0
 
 First release